changed the interface for opening a window in a direction. we now pass the direction value through the data parameter of the callback function, and as such we define a 'menu' program in config.h to be spawned when the keybind is handled. we set the open direction to the data and then use posix_spawnp on the config menu.
the format and scope of config has also been increased. i prepended all
of the global config constants with 'cfg_' because i think that makes
bonsai.c a lot more readable. toying around with doing this as a general
style rule, where functions/variables are prepended with '<tag>_' to
denote a sort of 'namespace' idea, while the rest of the identifier is
still in our typical lowercaseoneword form.